useInterval: move to hooks folder
